requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
boto3==1.34.123
botocore==1.34.162
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.121.2
h11==0.16.0
idna==3.11
jmespath==1.0.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9.1
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
tinydb==4.8.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
